fix(python2and3): use decode only on inline signature
When an inline signature is calculated using pyopenssl, the sign/base64
methods return a signature of type bytes in python 3, but when using the
non-inline method of getting a signature it returns a type of str.  We
need both to be str so decode only the inline signature.

Working on #20