fix(moimpl): avoid conflict between rn prefixes by not rstripping hypens
There are two use cases for using the rn prefix to find the proper child
class container:

1. Access child classes as if they were attributes (lookup)
2. Attach children to parent Mo's once the children are initialized.

If sibling mo's have nearly identical rn prefixes ('a' and 'a-' for example),
depending on the order in which rn prefixes are processed the wrong child
class container may be used.

Avoid this by not stripping hyphen off the rn prefix and on lookup add a
hyphen if the mo is a singleton (has no naming properties).

Working on #20