fix(python2and3): remove dependance on __cmp__ and fix MoStatus
In Python 3 __cmp__ is ignored and the cmp builtin is removed.  Replace
all __cmp__ methods with the 6 rich comparison methods.  Fix some
minor typos in the docs.  Fix MoStatus.fromString() and align the current
status with the model where the default value is 0 and 1 is not a valid
value.  Fix the tests for MoStatus.

Closes #46