chore(codec): add a comment about why sort_keys is used for JSON
The APIC REST API is not 100% json compliant when it comes to the JSON parser.
The APIC REST API JSON format is derived from the APIC REST API XML and
requires the object order to contain the attributes object before the
children array.  To accomplish this we use sort keys.