fix(accessimpl): handle redirects during login
For a redirected POST, requests follows what browsers do and redirects to the
new location but uses a GET during the redirect.  If the endpoint has redirect
enabled for http to https, this behavior can be problematic since our first
interaction with the endpoint may be a login request which is a POST.

When a redirect occurs during login the new location can be used to update the
session url so future interaction with the endpoint works as expected, assuming
the secure option is set properly by the user.

Closes #56